refactor: split scraper fetch.py (1751 lines) into 5 focused modules
Extract cache.py (Valkey cache + freshness revalidation),
proxy.py (httpx + Playwright proxy config), dns_guard.py
(DNS rebinding + SSRF protection), barrier.py (bot challenge
detection), and fetch_strategy.py (three-tier fetch pipeline).

fetch.py is now a thin re-export. Updated politeness.py and
recovery.py imports. Pure function-level copy — no behavioral
changes, no new dependencies, no Dockerfile changes needed
(COPY scraper-svc/scraper/ is directory-wide).

Closes #188
